Converted handshake packet & client status request packet
